mid spawn mobs

+++ core/src/main/java/com/mygdx/game/net/messages/EntityInfo.java
@@ -3,13 +3,13 @@ package com.mygdx.game.net.messages;
 import com.mygdx.game.gamestate.factories.MobsFactory;
 
 public class EntityInfo {
-    public float id;
+    public long id;
     public float x, y, xSpeed, ySpeed;
     public MobsFactory.Type type;
     public String name;
     public float hp;
 
-    public EntityInfo set(float id, String name, float hp, float x, float y, float xS, float yS){
+    public EntityInfo set(long id, String name, float hp, float x, float y, float xS, float yS){
         this.x = x;
         this.y = y;
         xSpeed = xS;
